functions/requirements.txt
firebase_functions~=0.1.

google-cloud-storage
google-cloud-videointelligence
